Use atexit to clean up after simulators
`close` is registered as an `atexit` method to help in cases such as
people Ctrl+C-ing the GUI.

+++ nengo_spinnaker/simulator.py
@@ -1,3 +1,4 @@
+import atexit
 import logging
 import nengo
 from nengo.cache import get_default_decoder_cache
@@ -47,6 +48,9 @@ class Simulator(object):
             Duration of one period of the simulator. This determines how much
             memory will be allocated to store precomputed and probed data.
         """
+        # Register `close` to be called when the interpreter is closed.
+        atexit.register(self.close)
+
         # Create a controller for the machine and boot if necessary
         hostname = rc.get("spinnaker_machine", "hostname")
         machine_width = rc.getint("spinnaker_machine", "width")
@@ -142,7 +146,11 @@ class Simulator(object):
 
     def __enter__(self):
         """Enter a context which will close the simulator when exited."""
-        pass
+        # Return self to allow usage like:
+        #
+        #     with nengo_spinnaker.Simulator(model) as sim:
+        #         sim.run(1.0)
+        return self
 
     def __exit__(self, exception_type, exception_value, traceback):
         """Exit a context and close the simulator."""
@@ -278,10 +286,11 @@ class Simulator(object):
 
     def close(self):
         """Clean the SpiNNaker board and prevent further simulation."""
-        # Stop the application
-        self._closed = True
-        self.io_controller.close()
-        self.controller.send_signal("stop")
+        if not self._closed:
+            # Stop the application
+            self._closed = True
+            self.io_controller.close()
+            self.controller.send_signal("stop")
 
     def trange(self, dt=None):
         return np.arange(1, self.steps + 1) * (self.dt or dt)
